+++ Codebase/equaly/src/main/java/org/hwr/equaly/controller/languageTagger/LanguageTagger.java
@@ -1,7 +1,7 @@
-package org.hwr.equaly.controller.languageDetector;
+package org.hwr.equaly.controller.languageTagger;
 
 import com.github.pemistahl.lingua.api.Language;
 
-public interface LanguageDetector {
+public interface LanguageTagger {
     Language getLanguage(String text);
 }
