+++ Codebase/equaly/src/main/java/org/hwr/equaly/controller/languageTagger/LanguageTaggerImpl.java
@@ -1,4 +1,4 @@
-package org.hwr.equaly.controller.languageDetector;
+package org.hwr.equaly.controller.languageTagger;
 
 import com.github.pemistahl.lingua.api.Language;
 import com.github.pemistahl.lingua.api.LanguageDetectorBuilder;
@@ -7,10 +7,10 @@ import static com.github.pemistahl.lingua.api.Language.*;
 import org.springframework.stereotype.Component;
 
 @Component
-public class LanguageDetectorImpl implements LanguageDetector {
+public class LanguageTaggerImpl implements LanguageTagger {
     private com.github.pemistahl.lingua.api.LanguageDetector detector;
 
-    public LanguageDetectorImpl() {
+    public LanguageTaggerImpl() {
         this.detector = LanguageDetectorBuilder.fromLanguages(ENGLISH, GERMAN).build();
     }
